# BackEnd/swagger.yaml
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
# Added by API Auto Mocking Plugin
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/MatteoLunardon/TRENTunOAPIs_la_vendetta/1.0.0
- url: https://virtserver.swaggerhub.com/JOAOPROGRAMMING/TRENTunO/1.0.0
  description: TRENTunO API Auto Mocking
...
paths:
...
  /reviews:
    get:
      tags:
      - Review
      summary: Retrieves reviews
      description: Returns all reviews or filters by user ID or park ID.
      parameters:
      - name: user_id
        in: query
        description: Filter reviews by the user ID
        required: false
        schema:
          type: string
      - name: park_id
        in: query
        description: Filter reviews by the park ID
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reviews'
        "400":
          description: Invalid input format
        "404":
          description: No reviews found
        "500":
          description: Server error

    post:
      tags:
      - Review
      summary: Adds a new review
      description: Creates a new review in the system.
      requestBody:
        description: The review object to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reviews'
      responses:
        "201":
          description: Review added successfully!
        "400":
          description: Invalid input format
  /reviews/{id}:
    get:
      tags:
      - Review
      summary: Retrieve a review by ID
      description: Returns the review with the specified ID.
      parameters:
      - name: id
        in: path
        description: The unique ID of the review
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reviews'
        "400":
          description: Invalid input format
        "404":
          description: Review not found
        "500":
          description: Server error

    delete:
      tags:
      - Review
      summary: Deletes a review by ID
      description: Removes the review with the specified ID from the system.
      parameters:
      - name: id
        in: path
        description: The unique ID of the review to delete
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Review successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Review successfully deleted
                  reviewDeleted:
                    $ref: '#/components/schemas/Reviews'
        "400":
          description: Invalid input format
        "404":
          description: Review not found
        "500":
          description: Server error
...
  /reports:
    get:
      tags:
      - Report
      description: Returns all reports in the system (filterable by `user_id` and `park_id`).
      parameters:
      - name: user_id
        in: query
        description: The user ID to filter reports by
        required: false
        schema:
          type: string
      - name: park_id
        in: query
        description: The park ID to filter reports by
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reports'
        "400":
          description: Invalid input format
        "404":
          description: Report not found
        "500":
          description: Server error

    post:
      tags:
      - Report
      description: Creates a new report in the system.
      requestBody:
        description: The details of the report to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reports'
      responses:
        "201":
          description: Successfully created a new report
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Report successfully created
                  report:
                    $ref: '#/components/schemas/Reports'
        "400":
          description: Invalid input format
        "500":
          description: Server error
  /reports/{id}:
    get:
      tags:
      - Report
      description: Returns a specific report by its unique ID.
      parameters:
      - name: id
        in: path
        description: The unique ID of the report
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reports'
        "400":
          description: Invalid input format
        "404":
          description: Report not found
        "500":
          description: Server error

    delete:
      tags:
      - Report
      description: Deletes a specific report by its unique ID.
      parameters:
      - name: id
        in: path
        description: The unique ID of the report to delete
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successfully deleted the report
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Report successfully deleted
        "400":
          description: Invalid input format
        "404":
          description: Report not found
        "500":
          description: Server error
components:
  schemas:
...
    Reports:
      type: object
      properties:
        id:
          type: string
          example: "60d21b4667d0d8992e610c85"
        user_id:
          type: string
          example: "60d21b4567d0d8992e610b97"
        park_id:
          type: string
          example: "60d21b4567d0d8992e610b98"
        status:
          type: boolean
          example: true
        description:
          type: string
          example: "Broken bench near the entrance"
    Reviews:
      type: object
      properties:
        id:
          type: string
          example: "61d21b4567d0d8992e610c12"
        user_id:
          type: string
          example: "62e31b4567d0d8992e610d34"
        park_id:
          type: string
          example: "60d21b4567d0d8992e610b98"
        rating:
          type: number
          example: 4.5
        comment:
          type: string
          example: "A lovely place to relax."
...